backend/requirements.txt
flask
flask_restful
mysqlclient
requests
# line-bot-sdk 
mysql-connector-python